Kernel+LibC: Fix various build issues introduced by ssize_t
Now that ssize_t is derived from size_t, we have to

+++ Kernel/Devices/FullDevice.cpp
@@ -47,7 +47,7 @@ bool FullDevice::can_read(const FileDescription&, size_t) const
 
 ssize_t FullDevice::read(FileDescription&, size_t, u8* buffer, ssize_t size)
 {
-    ssize_t count = min(PAGE_SIZE, size);
+    ssize_t count = min(static_cast<ssize_t>(PAGE_SIZE), size);
     memset(buffer, 0, (size_t)count);
     return count;
 }
